# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ti Saúde API",
        "description": "Documentação da API do desafio fullstack",
        "version": "1.0.1"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/especialidades": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Lista todas as especialidades",
                "operationId": "6584440eceb4e8121978fbb85682918b",
                "responses": {
                    "200": {
                        "description": "Especialidades listadas com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Cria uma nova especialidade",
                "operationId": "815ad5cf55785ef25a707dc07d1fdf1b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "espec_nome"
                                ],
                                "properties": {
                                    "espec_codigo": {
                                        "type": "string",
                                        "example": "CARD"
                                    },
                                    "espec_nome": {
                                        "type": "string",
                                        "example": "Cardiologia"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Especialidade criada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/especialidades/{id}": {
            "get": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Busca uma especialidade por ID",
                "operationId": "402e61363f39342d5f560324063917b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialidade encontrada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Especialidade não encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Atualiza uma especialidade",
                "operationId": "b0468a2dc51d9af3301f0eff5e987272",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "espec_nome"
                                ],
                                "properties": {
                                    "espec_codigo": {
                                        "type": "string",
                                        "example": "CARD"
                                    },
                                    "espec_nome": {
                                        "type": "string",
                                        "example": "Cardiologia Intervencionista"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Especialidade atualizada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Especialidade não encontrada"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Especialidades"
                ],
                "summary": "Remove uma especialidade",
                "operationId": "14e81c3265ef55591db4be821cd0c651",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Especialidade removida com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Especialidade não encontrada"
                    },
                    "409": {
                        "description": "Especialidade possui médicos vinculados"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}